+++ src/parallel/impl_par_methods.rs
@@ -5,6 +5,8 @@ use crate::parallel::prelude::*;
 use crate::parallel::par::ParallelSplits;
 use super::send_producer::SendProducer;
 
+use crate::partial::Partial;
+
 /// # Parallel methods
 ///
 /// These methods require crate feature `rayon`.
@@ -162,77 +164,4 @@ zip_impl! {
     [true P1 P2 P3 P4],
     [true P1 P2 P3 P4 P5],
     [false P1 P2 P3 P4 P5 P6],
-}
-
-/// Partial is a partially written contiguous slice of data;
-/// it is the owner of the elements, but not the allocation,
-/// and will drop the elements on drop.
-#[must_use]
-pub(crate) struct Partial<T> {
-    /// Data pointer
-    ptr: *mut T,
-    /// Current length
-    len: usize,
-}
-
-impl<T> Partial<T> {
-    /// Create an empty partial for this data pointer
-    ///
-    /// Safety: Unless ownership is released, the 
-    /// Partial acts as an owner of the slice of data (not the allocation);
-    /// and will free the elements on drop; the pointer must be dereferenceable
-    /// and the `len` elements following it valid.
-    pub(crate) unsafe fn new(ptr: *mut T) -> Self {
-        Self {
-            ptr,
-            len: 0,
-        }
-    }
-
-    pub(crate) fn stub() -> Self {
-        Self { len: 0, ptr: 0 as *mut _ }
-    }
-
-    pub(crate) fn is_stub(&self) -> bool {
-        self.ptr.is_null()
-    }
-
-    /// Release Partial's ownership of the written elements, and return the current length
-    pub(crate) fn release_ownership(mut self) -> usize {
-        let ret = self.len;
-        self.len = 0;
-        ret
-    }
-
-    /// Merge if they are in order (left to right) and contiguous.
-    /// Skips merge if T does not need drop.
-    pub(crate) fn try_merge(mut left: Self, right: Self) -> Self {
-        if !std::mem::needs_drop::<T>() {
-            return left;
-        }
-        // Merge the partial collect results; the final result will be a slice that
-        // covers the whole output.
-        if left.is_stub() {
-            right
-        } else if left.ptr.wrapping_add(left.len) == right.ptr {
-            left.len += right.release_ownership();
-            left
-        } else {
-            // failure to merge; this is a bug in collect, so we will never reach this
-            debug_assert!(false, "Partial: failure to merge left and right parts");
-            left
-        }
-    }
-}
-
-unsafe impl<T> Send for Partial<T> where T: Send { }
-
-impl<T> Drop for Partial<T> {
-    fn drop(&mut self) {
-        if !self.ptr.is_null() {
-            unsafe {
-                std::ptr::drop_in_place(std::slice::from_raw_parts_mut(self.ptr, self.len));
-            }
-        }
-    }
 }
